requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
boto3==1.24.74
botocore==1.27.74
certifi==2022.9.14
charset-normalizer==2.1.1
click==8.1.3
Django==4.1.1
django-cors-headers==3.13.0
django-environ==0.9.0
django-storages==1.13.1
djangorestframework==3.13.1
filelock==3.8.0
idna==3.4
jmespath==1.0.1
joblib==1.1.0
numpy==1.23.3
pandas==1.4.4
python-dateutil==2.8.2
pytz==2022.2.1
regex==2022.9.13
requests==2.28.1
s3transfer==0.6.0
sacremoses==0.0.53
sentencepiece==0.1.97
six==1.16.0
sqlparse==0.4.2
tokenizers==0.7.0
torch==1.10.1
tqdm==4.64.1
transformers==2.10.0
typing_extensions==4.3.0
urllib3==1.26.12
